StorageManager
The StorageManager class currently stores the following:

A copy of the  inputted image in a folder called "InputImages". The copy's name is the current "yyyyMMddHHmmss"
The text from the image in a string variable called "storedImageText"
The coordinates of the image text in a hash map called "storedWordMap"
The translated text in a string variable called "storedTranslation" and in a text file called the current "yyyyMMddHHmmss". The text file is then placed in a history folder.

If any of the file or folder names need to be changed let me know

Also the StorageManager is the only thing added to the program. No other classes were touched.